move NeuronGroupParams one level up
